Add MTK GED KPI support to fix broken Mediatek gpufreq
Mediatek GPU scheduler likes to have the timestamps of the frames to be
able to adjust DVFS.
Technically this isn't completely needed, because it looks like DVFS can
work once we started triggering /proc/ged. But now that the work is
done, let's do it completely.

In benchmarks on Poco X4 GT, before this patch result is 1500, after is
5700. If we disable the patch after enabling it without rebooting, it
goes to 5400. So looks like GED KPI thingy does try to do a bit more
than just standard DVFS.

Thanks @sarthakroy2002 for the help and support (other people helped as
well)

[xyzuniverse]: Trigger the GED if the hwcomposer is mtk_common.

Change-Id: Ic29a231ea8651efd598083611197aaa9e3c1fbbe
Signed-off-by: Ido <dev.xyzuniverse@gmail.com>