Optimize the priority of ServiceManager(1/2).
During the application startup, we find that the Servicemanager thread is frequently preempted
due to its low priority, resulting in slow application startup and responsiveness.
We hope to optimize these issues by adjusting the priority to RT(98) of the SM
to avoid time-consuming binder in clients to ServiceManager, and improve application startup and response speed.

Bug: 419835707

Change-Id: I58d6baed9a01b38ec804273063ced7a24d12b842
Signed-off-by: zhanghao56 <zhanghao56@xiaomi.corp-partner.google.com>
Signed-off-by: xyzuniverse <dev.xyzuniverse@gmail.com>